change multistatus to writestatus and add helpers
