+++ lib/tempoiq.js
@@ -7,7 +7,7 @@ var BulkWrite = require('./models/bulk_write');
 var Cursor = require('./models/cursor');
 var DataPoint = require('./models/datapoint');
 var Device = require('./models/device');
-var MultiStatus = require('./models/multi_status');
+var WriteStatus = require('./models/write_status');
 var Pipeline = require('./models/pipeline');
 var Query = require('./models/query');
 var Row = require('./models/row');
@@ -223,13 +223,13 @@ TempoIQ.Client = function(key, secret, host, opts) {
   // is generally useful for importing data to many devices at once.
   //
   // @param {BulkWrite} write BulkWrite with device / sensor data
-  // @param {function(err, multiStatus)} callback Returns a MultiStatus with the results of the write
+  // @param {function(err, multiStatus)} callback Returns a WriteStatus with the results of the write
   base.writeBulk = function(write, callback) {
     base._session.post("/v2/write", JSON.stringify(write), {}, callback, function(result) {
       if (result.code == HttpResult.OK) {
-        return { error: null, payload: new MultiStatus };
+        return { error: null, payload: new WriteStatus };
       } else if (result.code == HttpResult.MULTI) {
-        return { error: null, payload: new MultiStatus(JSON.parse(result.body)) };
+        return { error: null, payload: new WriteStatus(JSON.parse(result.body)) };
       }
     });
   };
